requirements.txt
pillow==2.9.0
requests==2.7.0
twython==3.2.0
yahoo-finance==1.2.1
